Add files via upload
A secure web application that takes an address entry as the input from a user and finds the zip code of that particular address. The entire application was implemented using JavaScript and jQuery libraries. The app is made to work for addresses throughout the world by the use of the latest version of the Google Geo-location API.