Add files via upload
This particular PHP code uploads the files selected by the user to a particular folder on a server. The code clearly mentions the target directory where the files should be uploaded onto, and it does not allow duplication, i.e., it doesn't allow the file to be uploaded if the file is already present in that folder. Proper error messages are displayed to handle such conditions. Options have been provided for the size limits and the types of files that can be uploaded.